added payment handlers

+++ common/payment_client/payment_client.go
@@ -2,6 +2,6 @@ package payment_client
 
 // StripeClient interface defines methods for interacting with Stripe.
 type PaymentClient interface {
-	CreateCheckoutSession(orderID string, totalAmount int64) (string, error)
+	CreateCheckoutSession(orderID string, totalAmount float64) (string, error)
 	HandleWebhook(payload []byte, sigHeader string) (*PaymentIntentRes, error)
 }
